Convert LLVM compiler to TypeScript
