+++ src/command-native.js
@@ -10,7 +10,7 @@ var fs            = require('fs'),
 
 // Load source map support and LLVM target
 require('source-map-support').install()
-require('./targets/llvm')
+var LLVMCompiler = require('./targets/llvm').LLVMCompiler
 
 // Command-line arguments ----------------------------------------------------
 
@@ -112,8 +112,9 @@ try {
   process.exit(1)
 }
 
-var outputs = []
-tree.emitToFile({logger: logger, outputs: outputs, dump: Opts.dump})
+var outputs  = [],
+    compiler = new LLVMCompiler()
+compiler.emitToFile(tree, {logger: logger, outputs: outputs, dump: Opts.dump})
 
 function objectFileForBitcodeFile (path) {
   return path.replace(/\.bc$/, '.o')
